+++ packages/terra-toggle-button/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Changed
+  * Replaced `terra-button` with `IconButton`
   * Updated jest snapshots.
   
 ## 3.61.3 - (May 3, 2022)