+++ packages/terra-toggle-button/CHANGELOG.md
@@ -6,6 +6,11 @@
   * Replaced `terra-button` with `IconButton`
   * Updated jest snapshots.
   
+## 3.61.4 - (July 5, 2022)
+
+* Changed
+  * Minor dependency version bump
+
 ## 3.61.3 - (May 3, 2022)
 
 * Changed